+++ .storybook/webpack.config.js
@@ -1,13 +1,14 @@
-const path = require('path')
-module.exports = ({config, mode}) => {
+module.exports = ({config}) => {
   config.module.rules.push({
     test: /\.(ts|tsx)$/,
-    use: [
-      {
-        loader: require.resolve('ts-loader'),
-      },
-    ],
+    loader: 'babel-loader',
+    options: {
+      customize: require.resolve('babel-preset-react-app/webpack-overrides'),
+      presets: [['react-app', {flow: false, typescript: true}]],
+    },
   })
+
   config.resolve.extensions.push('.ts', '.tsx')
+
   return config
 }