+++ index.ts
@@ -1,3 +1,0 @@
-export {drawQRCode} from './src/drawQRCode'
-export {renderQRCode} from './src/renderQRCode'
-export * from './src/types'