+++ src/index.ts
@@ -1,0 +1,3 @@
+export {drawQRCode} from './drawQRCode'
+export {renderQRCode} from './renderQRCode'
+export * from './types'