+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# 2.0.0
+
+**Improvements**:
+
+- Improve how padding is calculated and displayed
+- Optimize the Bloom Logo
+- Remove ID generation
+- Use tsdx for building
+
+**Breaking Changes**:
+
+- Omit `canvas` from the update function
+- Simplify typing
+  - `DrawConfig`/`RenderConfig` => `Config`
+  - `DrawResult`/`RenderResult` => `Result`
+
 # 1.0.0
 
 - Initial Release